mejoras cuenta

+++ app/Http/Controllers/BancosController.php
@@ -4,9 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Banco;
+use Illuminate\Database\QueryException;
 
 use App\Http\Requests;
 
+/*
+ * 'codigo' => 'A01' , 'accion' => 'BancoController@show'
+ * 'codigo' => 'A02' , 'accion' => 'BancoController@delete'
+ * 'codigo' => 'A03' , 'accion' => 'BancoController@update'
+ * 'codigo' => 'A04' , 'accion' => 'BancoController@create'
+ * 'codigo' => 'A05' , 'accion' => 'BancoController@index'
+*/
 class BancosController extends Controller
 {
  
@@ -17,6 +25,7 @@ class BancosController extends Controller
      */
     public function index()
     {
+      $this->authorize('A05');
       $bancos = Banco::orderBy('id', 'desc')->paginate(10);
       return view('bancos.index', compact('bancos'));
     }
@@ -27,7 +36,8 @@ class BancosController extends Controller
      * @return Response
      */
     public function create()
-    {
+    { 
+      $this->authorize('A04');
       return view('bancos.create');
     }
 
@@ -39,6 +49,7 @@ class BancosController extends Controller
      */
     public function store(Request $request)
     {
+      $this->authorize('A04');
       $validator = Banco::validador($request->all());
       if ($validator->fails()) {
         return redirect('bancos/create')
@@ -47,7 +58,7 @@ class BancosController extends Controller
       }
       else {
         $banco =Banco::create($request->all());
-        return redirect()->route('bancos.show',['id' => $banco->id ])->with('message', 'Item created successfully.');
+        return redirect()->route('bancos.show',['id' => $banco->id ])->with('success', 'Banco correctamente creado.');
       }
     } 
 
@@ -60,6 +71,7 @@ class BancosController extends Controller
      */
     public function show($id)
     {
+      $this->authorize('A01');
       $banco = Banco::findOrFail($id);
       return view('bancos.show', compact('banco'));
     }
@@ -72,6 +84,7 @@ class BancosController extends Controller
      */
     public function edit($id)
     {
+      $this->authorize('A03');
       $banco = Banco::findOrFail($id);
       return view('bancos.edit', compact('banco'));
     }
@@ -85,6 +98,7 @@ class BancosController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('A03');
         $banco = Banco::findOrFail($id);
         $values = $request->all(); 
         $validator = Banco::validador($values,$banco);
@@ -95,7 +109,7 @@ class BancosController extends Controller
             ->withInput();
         }
         else {
-          return redirect()->route('bancos.show',['id' => $banco->id ])->with('message', 'Item update successfully.');
+          return redirect()->route('bancos.show',['id' => $banco->id ])->with('success', 'Banco correctamente actualizado');
         }
     }
 
@@ -107,9 +121,14 @@ class BancosController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('A02');
         $banco = Banco::findOrFail($id);
-        $banco->delete();
-
-        return redirect()->route('bancos.index')->with('message', 'Item deleted successfully.');
+        try {
+            $banco->delete();
+        }catch (QueryException $e){
+            return redirect()->route('bancos.index')->with('danger', 'Este Banco esta siendo utilizado.');
+        }
+        return redirect()->route('bancos.index')->with('success', 'Banco Elminado.');
     }
+
 }
